requirements.txt
requests>=2.28.0
PyYAML>=6.0
dnspython>=2.4.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
